add Service classes, add javadoc for all current methods

+++ src/main/java/com/hxngxd/entities/Book.java
@@ -1,14 +1,17 @@
 package com.hxngxd.entities;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
+/**
+ * Lớp Book đại diện cho một cuốn sách trong thư viện.
+ */
 public class Book {
+
     private int id;
     private String title;
-    private int yearOfPublication;
+    private short yearOfPublication;
     private String shortDescription;
-    // private Img coverImage;
+//    private Img coverImage;
     private int numberOfPages;
     private LocalDateTime dateAdded;
     private LocalDateTime lastUpdated;
@@ -16,7 +19,13 @@ public class Book {
     private int totalCopies;
     private double averageRating;
 
-    public double calculateAverageRating(){
-        return 0;
+    /**
+     * Tính điểm đánh giá trung bình.
+     *
+     * @return Điểm đánh giá trung bình.
+     */
+    public double calculateAverageRating() {
+        return 0.0;
     }
+
 }
